+++ src/main/java/br/com/zalf/prolog/webservice/cs/nps/PesquisaNpsDao.java
@@ -3,6 +3,7 @@ package br.com.zalf.prolog.webservice.cs.nps;
 import br.com.zalf.prolog.webservice.cs.nps.model.PesquisaNpsBloqueio;
 import br.com.zalf.prolog.webservice.cs.nps.model.PesquisaNpsDisponivel;
 import br.com.zalf.prolog.webservice.cs.nps.model.PesquisaNpsRealizada;
+import br.com.zalf.prolog.webservice.frota.veiculo.historico._model.OrigemAcaoEnum;
 import org.jetbrains.annotations.NotNull;
 
 import java.util.Optional;
@@ -17,7 +18,9 @@ public interface PesquisaNpsDao {
     Optional<PesquisaNpsDisponivel> getPesquisaNpsColaborador(@NotNull final Long codColaborador) throws Throwable;
 
     @NotNull
-    Long insereRespostasPesquisaNps(@NotNull final PesquisaNpsRealizada pesquisaRealizada) throws Throwable;
+    Long insereRespostasPesquisaNps(@NotNull final OrigemAcaoEnum origemResposta,
+                                    @NotNull final PesquisaNpsRealizada pesquisaRealizada) throws Throwable;
 
-    void bloqueiaPesquisaNpsColaborador(@NotNull final PesquisaNpsBloqueio pesquisaBloqueio) throws Throwable;
+    void bloqueiaPesquisaNpsColaborador(@NotNull final OrigemAcaoEnum origemBloqueio,
+                                        @NotNull final PesquisaNpsBloqueio pesquisaBloqueio) throws Throwable;
 }