+++ src/main/java/br/com/zalf/prolog/webservice/cs/nps/PesquisaNpsDaoImpl.java
@@ -5,6 +5,7 @@ import br.com.zalf.prolog.webservice.cs.nps.model.PesquisaNpsBloqueio;
 import br.com.zalf.prolog.webservice.cs.nps.model.PesquisaNpsDisponivel;
 import br.com.zalf.prolog.webservice.cs.nps.model.PesquisaNpsRealizada;
 import br.com.zalf.prolog.webservice.database.DatabaseConnection;
+import br.com.zalf.prolog.webservice.frota.veiculo.historico._model.OrigemAcaoEnum;
 import org.jetbrains.annotations.NotNull;
 
 import java.sql.Connection;
@@ -54,24 +55,27 @@ public final class PesquisaNpsDaoImpl extends DatabaseConnection implements Pesq
 
     @NotNull
     @Override
-    public Long insereRespostasPesquisaNps(@NotNull final PesquisaNpsRealizada pesquisaRealizada)
+    public Long insereRespostasPesquisaNps(@NotNull final OrigemAcaoEnum origemResposta,
+                                           @NotNull final PesquisaNpsRealizada pesquisaRealizada)
             throws Throwable {
         Connection conn = null;
         PreparedStatement stmt = null;
         ResultSet rSet = null;
         try {
             conn = getConnection();
-            stmt = conn.prepareCall("{CALL CS.FUNC_NPS_INSERE_RESPOSTAS_PESQUISA(" +
-                    "F_COD_PESQUISA_NPS              := ?," +
-                    "F_COD_COLABORADOR_REALIZACAO    := ?," +
-                    "F_DATA_HORA_REALIZACAO_PESQUISA := ?," +
-                    "F_RESPOSTA_PERGUNTA_ESCALA      := ?," +
-                    "F_RESPOSTA_PERGUNTA_DESCRITIVA  := ?)}");
+            stmt = conn.prepareCall("{call cs.func_nps_insere_respostas_pesquisa(" +
+                                            "f_cod_pesquisa_nps              := ?," +
+                                            "f_cod_colaborador_realizacao    := ?," +
+                                            "f_data_hora_realizacao_pesquisa := ?," +
+                                            "f_resposta_pergunta_escala      := ?," +
+                                            "f_resposta_pergunta_descritiva  := ?," +
+                                            "f_origem_resposta               := ?)}");
             stmt.setLong(1, pesquisaRealizada.getCodPesquisaNps());
             stmt.setLong(2, pesquisaRealizada.getCodColaboradorRealizacao());
             stmt.setObject(3, Now.getOffsetDateTimeUtc());
             stmt.setShort(4, pesquisaRealizada.getRespostaPerguntaEscala());
             stmt.setString(5, pesquisaRealizada.getRespostaPerguntaDescritiva());
+            stmt.setString(6, origemResposta.asString());
             rSet = stmt.executeQuery();
             if (rSet.next() && rSet.getLong(1) != 0) {
                 return rSet.getLong(1);
@@ -84,18 +88,21 @@ public final class PesquisaNpsDaoImpl extends DatabaseConnection implements Pesq
     }
 
     @Override
-    public void bloqueiaPesquisaNpsColaborador(final @NotNull PesquisaNpsBloqueio pesquisaBloqueio) throws Throwable {
+    public void bloqueiaPesquisaNpsColaborador(@NotNull final OrigemAcaoEnum origemBloqueio,
+                                               @NotNull final PesquisaNpsBloqueio pesquisaBloqueio) throws Throwable {
         Connection conn = null;
         PreparedStatement stmt = null;
         try {
             conn = getConnection();
-            stmt = conn.prepareCall("{CALL CS.FUNC_NPS_BLOQUEIA_PESQUISA(" +
-                    "F_COD_PESQUISA_NPS            := ?," +
-                    "F_COD_COLABORADOR_BLOQUEIO    := ?," +
-                    "F_DATA_HORA_BLOQUEIO_PESQUISA := ?)}");
+            stmt = conn.prepareCall("{call cs.func_nps_bloqueia_pesquisa(" +
+                                            "f_cod_pesquisa_nps            := ?," +
+                                            "f_cod_colaborador_bloqueio    := ?," +
+                                            "f_data_hora_bloqueio_pesquisa := ?," +
+                                            "f_origem_bloqueio_pesquisa    := ?)}");
             stmt.setLong(1, pesquisaBloqueio.getCodPesquisaNps());
             stmt.setLong(2, pesquisaBloqueio.getCodColaboradorBloqueio());
             stmt.setObject(3, Now.getOffsetDateTimeUtc());
+            stmt.setString(4, origemBloqueio.asString());
             stmt.execute();
         } finally {
             close(conn, stmt);